Leap integrated operating system step one >>
this file contains several examples and the of the the leap SDK, its a
quick and dirty push as it stands as of January 14 2014. It will likely
provide a decent starting point for anyone working with C# leap and
unity, and my intension is to move into the next phase of the project
building on Archlinux and Fluxbox to build a gesture centric operating
system for AR.

Special thanks to Leap motion for the incredible technology.